+++ docs/changelog.md
@@ -1,7 +1,13 @@
 # TBD
 
-# 0.1.1
+# 0.1.2
 
+### Features
+* Defined Lambda Service GRPC contract
+* Add `regenerate-protobuf-output` script file which uses the latest version of `generate-protobuf-bindings` from the devtools repo to generate bindings files
+* Generated the Golang and Typescript bindings files
+
+# 0.1.1
 * Init commit
 * Add base directories and files
   * Add changelog file